+++ fbw-a32nx/src/systems/instruments/src/EFB/index.tsx
@@ -6,6 +6,13 @@ import React from 'react';
 import { render } from '@instruments/common/index';
 import { EfbWrapper } from '@flybywiresim/flypad';
 import { A320FailureDefinitions } from '@failures';
+import { EventBusContextProvider } from '../../../../../../fbw-common/src/systems/instruments/src/EFB/event-bus-provider';
 
 // TODO: Move failure definition into VFS
-render(<EfbWrapper failures={A320FailureDefinitions} />, true, true);
+render(
+  <EventBusContextProvider>
+    <EfbWrapper failures={A320FailureDefinitions} />
+  </EventBusContextProvider>,
+  true,
+  true,
+);