+++ fbw-common/src/systems/instruments/src/EFB/Dashboard/Widgets/Reminders/MaintenanceReminder.tsx
@@ -27,16 +27,15 @@ const ActiveFailureCard: FC<ActiveFailureCardProps> = ({ ata, name }) => {
       onClick={() => {
         dispatch(setSearchQuery(name.toUpperCase()));
 
-        const lastFailurePath = findLatestSeenPathname(history, '/failures');
-
+        const lastFailurePath = findLatestSeenPathname(history, '/failures/failure-list');
         if (!ata) {
-          history.push('/failures/compact');
+          history.push('/failures/failure-list/compact');
         }
 
         if (!lastFailurePath || lastFailurePath.includes('comfort')) {
-          history.push(`/failures/comfort/${ata}`);
+          history.push(`/failures/failure-list/comfort/${ata}`);
         } else {
-          history.push('/failures/compact');
+          history.push('/failures/failure-list/compact');
         }
       }}
     >
@@ -51,7 +50,10 @@ export const MaintenanceReminder = () => {
   const { allFailures, activeFailures } = useFailuresOrchestrator();
 
   return (
-    <RemindersSection title={t('Dashboard.ImportantInformation.Maintenance.Title')} pageLinkPath="/failures">
+    <RemindersSection
+      title={t('Dashboard.ImportantInformation.Maintenance.Title')}
+      pageLinkPath="/failures/failure-list"
+    >
       <div className="flex flex-row flex-wrap">
         {Array.from(activeFailures)
           // Sorts the failures by name length, greatest to least