+++ fbw-common/src/systems/shared/src/failures/failures-orchestrator.spec.ts
@@ -1,4 +1,8 @@
-import { FailuresOrchestrator } from '.';
+// Copyright (c) 2021-2023 FlyByWire Simulations
+//
+// SPDX-License-Identifier: GPL-3.0
+
+import { FailuresOrchestrator } from './failures-orchestrator';
 import { getActivateFailureSimVarName, getDeactivateFailureSimVarName } from './sim-vars';
 import { flushPromises } from './test-functions';
 
@@ -86,7 +90,7 @@ const identifier = 123;
 const name = 'test';
 
 function orchestrator() {
-  return new FailuresOrchestrator(prefix, [[0, identifier, name]]);
+  return new FailuresOrchestrator(null, prefix, [[0, identifier, name]]);
 }
 
 function activateFailure(o: FailuresOrchestrator) {