+++ .github/CHANGELOG.md
@@ -92,6 +92,8 @@
 1. [EFCS] Fix ground spoiler retraction after increasing TLA slightly above 0 - @lukecologne (luke)
 1. [FWC] Improved LDG LT memo to take into account light position - @BravoMike99 (bruno_pt99)
 1. [PRESS] Add pressurization system failures - @mjuhe (Miquel Juhe)
+1. [EFB] Random Failures - @Maverickwoe (Garoomf)
+
 
 ## 0.11.0